backend/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
requests
google-auth
Flask-Cors
axios
flask-migrate